move retrieval to VeraDnssecChain
